+++ app/(secure)/(tabs)/earnings.tsx
