+++ app/(secure)/viewroom.tsx
