+++ hooks/AppProvider.tsx
@@ -1,32 +1,17 @@
-import { Colors } from "@/constants/Colors";
-import React, { useEffect } from "react";
-import { ActivityIndicator, View } from "react-native";
-import useAuthStore from "../store/authStore";
+import React from "react";
 
 export default function AuthProvider({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const { checkAuthStatus } = useAuthStore();
-  const [isChecking, setIsChecking] = React.useState(true);
-
-  useEffect(() => {
-    const checkAuth = async () => {
-      await checkAuthStatus();
-      setIsChecking(false);
-    };
-
-    checkAuth();
-  }, []);
-
-  if (isChecking) {
-    return (
-      <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-        <ActivityIndicator size="large" color={Colors.primary} />
-      </View>
-    );
-  }
+  // if (isChecking) {
+  //   return (
+  //     <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+  //       <ActivityIndicator size="large" color={Colors.primary} />
+  //     </View>
+  //   );
+  // }
 
   return <>{children}</>;
 }
